refactor: refactor custom dialog to modal component
