+++ feat/time-capsule-detail/src/main/java/com/emotionstorage/time_capsule_detail/ui/modal/SaveChangesModal.kt
@@ -1,7 +1,12 @@
 package com.emotionstorage.time_capsule_detail.ui.modal
 
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import com.emotionstorage.ui.component.Modal
+import java.time.LocalDate
 
 @Composable
 fun SaveChangesModal(
@@ -22,4 +27,15 @@ fun SaveChangesModal(
             onDismiss = onDismiss,
         )
     }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun SaveChangesModalPreview() {
+    // background ui
+    Box(modifier = Modifier.fillMaxSize())
+
+    SaveChangesModal(
+        isModalOpen = true,
+    )
 }
